Moved code that could have been moved to partial handwritten files
