+++ src/SharpMetal/Handwritten/CGColorSpace.cs
